move service instance and rename mapping file
